Handle leading-dot localhost cookie aliases
